+++ logger.go
@@ -3,94 +3,8 @@ package log
 import (
 	"context"
 	"fmt"
-	. "github.com/logrusorgru/aurora"
-	"io"
 	"log"
-	"os"
 )
-
-type Level string
-
-var Constructor = NewLog(FileDepth(2))
-
-var StdLogger = Constructor.Log(FileDepth(3))
-
-var PrefixedStdLogger = Constructor.PrefixedLog(FileDepth(3))
-
-func Fatal(message interface{}, params ...interface{}) {
-	StdLogger.Fatal(message, params...)
-}
-
-func Error(message interface{}, params ...interface{}) {
-	StdLogger.Error(message, params...)
-}
-
-func Warn(message interface{}, params ...interface{}) {
-	StdLogger.Warn(message, params...)
-}
-
-func Debug(message interface{}, params ...interface{}) {
-	StdLogger.Debug(message, params...)
-}
-
-func Info(message interface{}, params ...interface{}) {
-	StdLogger.Info(message, params...)
-}
-
-func Trace(message interface{}, params ...interface{}) {
-	StdLogger.Trace(message, params...)
-}
-
-func FatalContext(ctx context.Context, message interface{}, params ...interface{}) {
-	StdLogger.FatalContext(ctx, message, params...)
-}
-
-func ErrorContext(ctx context.Context, message interface{}, params ...interface{}) {
-	StdLogger.ErrorContext(ctx, message, params...)
-}
-
-func WarnContext(ctx context.Context, message interface{}, params ...interface{}) {
-	StdLogger.WarnContext(ctx, message, params...)
-}
-
-func DebugContext(ctx context.Context, message interface{}, params ...interface{}) {
-	StdLogger.DebugContext(ctx, message, params...)
-}
-
-func InfoContext(ctx context.Context, message interface{}, params ...interface{}) {
-	StdLogger.InfoContext(ctx, message, params...)
-}
-
-func TraceContext(ctx context.Context, message interface{}, params ...interface{}) {
-	StdLogger.TraceContext(ctx, message, params...)
-}
-
-const (
-	FATAL Level = `FATAL`
-	ERROR Level = `ERROR`
-	WARN  Level = `WARN`
-	INFO  Level = `INFO`
-	DEBUG Level = `DEBUG`
-	TRACE Level = `TRACE`
-)
-
-var logColors = map[Level]string{
-	FATAL: BgRed(`[FATAL]`).String(),
-	ERROR: BgRed(`[ERROR]`).String(),
-	WARN:  BgYellow(`[ WARN]`).String(),
-	INFO:  BgBlue(`[ INFO]`).String(),
-	DEBUG: BgCyan(`[DEBUG]`).String(),
-	TRACE: BgMagenta(`[TRACE]`).String(),
-}
-
-var logTypes = map[Level]int{
-	FATAL: 0,
-	ERROR: 1,
-	WARN:  2,
-	INFO:  3,
-	DEBUG: 4,
-	TRACE: 5,
-}
 
 type SimpleLogger interface {
 	Print(v ...interface{})
@@ -156,86 +70,7 @@ func NewLog(options ...Option) Log {
 	}
 }
 
-type logOptions struct {
-	prefix    string
-	suffix    string
-	colors    bool
-	logLevel  Level
-	filePath  bool
-	fileDepth int
-	writer    io.Writer
-}
-
-func (lOpts *logOptions) applyDefault() {
-	lOpts.fileDepth = 2
-	lOpts.colors = true
-	lOpts.logLevel = TRACE
-	lOpts.filePath = true
-	lOpts.writer = os.Stdout
-}
-
-func (lOpts *logOptions) copy() *logOptions {
-	return &logOptions{
-		prefix:    lOpts.prefix,
-		suffix:    lOpts.suffix,
-		fileDepth: lOpts.fileDepth,
-		colors:    lOpts.colors,
-		logLevel:  lOpts.logLevel,
-		filePath:  lOpts.filePath,
-		writer:    lOpts.writer,
-	}
-}
-
-func (lOpts *logOptions) apply(options ...Option) {
-	for _, opt := range options {
-		opt(lOpts)
-	}
-}
-
-type Option func(*logOptions)
-
-func FileDepth(d int) Option {
-	return func(opts *logOptions) {
-		opts.fileDepth = d
-	}
-}
-
-func WithStdOut(w io.Writer) Option {
-	return func(opts *logOptions) {
-		opts.writer = w
-	}
-}
-
-func WithFilePath(enabled bool) Option {
-	return func(opts *logOptions) {
-		opts.filePath = enabled
-	}
-}
-
-func Prefixed(prefix string) Option {
-	return func(opts *logOptions) {
-		if opts.prefix != `` {
-			opts.prefix = fmt.Sprintf(`%s.%s`, opts.prefix, prefix)
-			return
-		}
-		opts.prefix = prefix
-	}
-}
-
-func WithColors(enabled bool) Option {
-	return func(opts *logOptions) {
-		opts.colors = enabled
-	}
-}
-
-func WithLevel(level Level) Option {
-	return func(opts *logOptions) {
-		opts.logLevel = level
-	}
-}
-
 func (l *logIpml) Log(options ...Option) Logger {
-
 	opts := l.logOptions.copy()
 	opts.apply(options...)
 
@@ -265,6 +100,30 @@ func (l *logIpml) PrefixedLog(options ...Option) PrefixedLogger {
 
 type logger struct {
 	logParser
+}
+
+func (l *logger) NewLog(opts ...Option) Logger {
+	defaults := l.logOptions.copy()
+	defaults.apply(opts...)
+
+	return &logger{
+		logParser: logParser{
+			logOptions: defaults,
+			log:        l.log,
+		},
+	}
+}
+
+func (l *logger) NewPrefixedLog(opts ...Option) PrefixedLogger {
+	defaults := l.logOptions.copy()
+	defaults.apply(opts...)
+
+	return &prefixedLogger{
+		logParser: logParser{
+			logOptions: defaults,
+			log:        l.log,
+		},
+	}
 }
 
 func (l *logger) ErrorContext(ctx context.Context, message interface{}, params ...interface{}) {
@@ -329,28 +188,4 @@ func (l *logger) Printf(format string, v ...interface{}) {
 
 func (l *logger) Println(v ...interface{}) {
 	l.logEntry(INFO, nil, l.WithPrefix(``, v), `INFO`)
-}
-
-func (l *logger) NewLog(opts ...Option) Logger {
-	defaults := l.logOptions.copy()
-	defaults.apply(opts...)
-
-	return &logger{
-		logParser: logParser{
-			logOptions: defaults,
-			log:        l.log,
-		},
-	}
-}
-
-func (l *logger) NewPrefixedLog(opts ...Option) PrefixedLogger {
-	defaults := l.logOptions.copy()
-	defaults.apply(opts...)
-
-	return &prefixedLogger{
-		logParser: logParser{
-			logOptions: defaults,
-			log:        l.log,
-		},
-	}
 }
