+++ noop_logger.go
