+++ prefixed_logger.go
@@ -9,6 +9,30 @@ type prefixedLogger struct {
 	logParser
 }
 
+func (l *prefixedLogger) NewLog(opts ...Option) Logger {
+	defaults := l.logOptions.copy()
+	defaults.apply(opts...)
+
+	return &logger{
+		logParser: logParser{
+			logOptions: defaults,
+			log:        l.log,
+		},
+	}
+}
+
+func (l *prefixedLogger) NewPrefixedLog(opts ...Option) PrefixedLogger {
+	defaults := l.logOptions.copy()
+	defaults.apply(opts...)
+
+	return &prefixedLogger{
+		logParser: logParser{
+			logOptions: defaults,
+			log:        l.log,
+		},
+	}
+}
+
 func (l *prefixedLogger) ErrorContext(ctx context.Context, prefix string, message interface{}, params ...interface{}) {
 	l.logEntry(ERROR, ctx, l.WithPrefix(prefix, message), params...)
 }
@@ -71,28 +95,4 @@ func (l *prefixedLogger) Printf(format string, v ...interface{}) {
 
 func (l *prefixedLogger) Println(v ...interface{}) {
 	l.logEntry(INFO, nil, v, l.colored(`INFO`))
-}
-
-func (l *prefixedLogger) NewLog(opts ...Option) Logger {
-	defaults := l.logOptions.copy()
-	defaults.apply(opts...)
-
-	return &logger{
-		logParser: logParser{
-			logOptions: defaults,
-			log:        l.log,
-		},
-	}
-}
-
-func (l *prefixedLogger) NewPrefixedLog(opts ...Option) PrefixedLogger {
-	defaults := l.logOptions.copy()
-	defaults.apply(opts...)
-
-	return &prefixedLogger{
-		logParser: logParser{
-			logOptions: defaults,
-			log:        l.log,
-		},
-	}
 }
